This example writes agents to files.
It creates a group of the agent name, and attached an attribute for each variable array

This doesn't support multiple iterations.

Need a better understanding of how to layout HL5 data to achieve that.

Running requires copying dlls to output dir.

There's a definite build issue, loads of warnings (field of class type without a DLL interface used in a class with a DLL interface), and some exceptions fail to output any text (presumably due to this), debugger said something about invalid string.